requirements.txt
pillow
qrcode
io